pyproject.toml
[tool.setuptools.packages.find]
include = ["server*"]

[project]
name = "m3u-studio"
version = "0.1.0"
description = "Local M3U playlist editor with drag-and-drop, live preview, and one-click export"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "httpx>=0.27",
    "pydantic>=2.9",
    "python-multipart>=0.0.9",
    "openai>=1.50",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.7",
    "pytest>=8.3",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "SIM"]
